requirements.txt
appnope==0.1.3
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.1
Django==2.2
django-extensions==3.1.5
django-filter==2.1.0
djangorestframework==3.13.1
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
Pygments==2.14.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.5
